Allow admin user to create runtime
We should allow the admin user (users with admin role)
to create runtime on the dashboard.

This is equivalent function to following CLI command

openstack runtime create $IMAGE --name python3,

The $IMAGE is a docker image like
openstackqinling/python3-runtime:0.0.2
which should be provided by the user.

According to qinling team, For now,
there is no need to do validation for that
input but in future the qinling-api should do some
sanity checks and reply to the UI accordingly.

Change-Id: I279442ac9b20dce3f141beffbfda7e8b5ef9c692
Story: 2004391
Task: 28018